Fix tests
