+++ src/Chapter03.Tests/Table03.03.Tests.cs
@@ -1,4 +1,3 @@
-using AddisonWesley.Michaelis.EssentialCSharp.Shared;
 using AddisonWesley.Michaelis.EssentialCSharp.Shared.Tests;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -9,18 +8,16 @@ public class ArrayHighlightsTests
 {
     public TestContext TestContext { get; set; } = null!; // Set by MSTest;
 
-    private const string FileName = "Table03.03.CommonArrayCodingErrors.cs";
-
     [TestMethod]
-    /* 1. */ [DataRow(nameof(CommonArayCodingErrors.SquareBracketsOnVariableRatherThanType), "CS0650")]
-    /* 2. */ [DataRow(nameof(CommonArayCodingErrors.NewKeywordWithDataTypeRequired), "CS1525", "CS1002", "CS1002", "CS1513", "CS1002", "CS1513", "CS1002")]
-    /* 3. */ [DataRow(nameof(CommonArayCodingErrors.ArraySizeCannotBeSpecifiedInDataType), "CS0270")]
-    /* 4. */ [DataRow(nameof(CommonArayCodingErrors.ArraySizeOrInitializerIsRequired), "CS1586")]
-    /* 5. */ [DataRow(nameof(CommonArayCodingErrors.ArraySizeWithEmptyInitializer), "CS0847")]
-    /* 9. */ [DataRow(nameof(CommonArayCodingErrors.MultiDimensionalArrayWithInconsistentSize), "CS0847")]
-    public async Task ParseAndCompile(string targetMethod, params string[] errorIds)
+    /* 1. */ [DataRow(".a", "CS0650")]
+    /* 2. */ [DataRow(".b", "CS1525", "CS1002", "CS1002", "CS1513", "CS1002", "CS1513", "CS1002")]
+    /* 3. */ [DataRow(".c", "CS0270")]
+    /* 4. */ [DataRow(".d", "CS1586")]
+    /* 5. */ [DataRow(".e", "CS0847")]
+    /* 9. */ [DataRow(".i", "CS0847")]
+    public async Task ParseAndCompile(string fileNameSuffix, params string[] errorIds)
     {
-        await CompilerAssert.CompileAsync("Table03.03.CommonArrayCodingErrors.cs", errorIds);
+        await CompilerAssert.CompileAsync($"Table03.03{fileNameSuffix}.CommonArrayCodingErrors.cs", errorIds);
     }
 
     // 6.
