+++ src/Chapter03/Table03.03.g.CommonArrayCodingErrors.cs
@@ -1,0 +1,16 @@
+﻿// Justification: Only snippets of source code shown for elucidation.
+#pragma warning disable CS0168 // Variable is declared but never used
+
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter03.Table03_03
+{
+    public partial class CommonArayCodingErrors
+    {
+        // 7.
+        static public void Hat0IsOnePastTheEndOfTheArray1()
+        {
+            int[] numbers =
+              new int[3];
+            numbers[^0] = 42;
+        }
+    }
+}
