+++ src/Chapter09.Tests/Chapter09.Tests.csproj
@@ -4,9 +4,6 @@
   </PropertyGroup>
   <Import Project="..\ChapterTests.props" />
   <ItemGroup>
-    <Compile Include="..\Shared\Tests\CompilerAssert.cs" Link="CompilerAssert.cs" />
-  </ItemGroup>
-  <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.4.0" />
   </ItemGroup>
   <ItemGroup>
